refactor helpers, add JSDocs
