+++ src/index.ts
@@ -3,4 +3,5 @@ export * from './platform/mock-local-home-platform';
 export * from './platform/stub-setup';
 export * from './platform/smart-home-app';
 export * from './platform/device-manager';
+export * from './platform/execute';
 (global as any).smarthome = smarthomeStub;
