+++ src/platform/execute.ts
@@ -1,3 +1,8 @@
+/**
+ * Holds Execute stubs and helpers
+ */
+/// <reference types="@google/local-home-sdk" />
+
 export const ExecuteStub: typeof smarthome.Execute = {
   Response: {
     Builder: class {
@@ -37,3 +42,34 @@ export const ExecuteStub: typeof smarthome.Execute = {
     },
   },
 };
+
+/**
+ * A helper to build a simple `ExecuteRequestCommands` for sending
+ * a single command to a single device
+ * @param deviceId  The id of the single device to send the command
+ * @param command  The single command to send to the device
+ * @param params  Parmeters for the command
+ */
+export function createSimpleExecuteCommands(
+  deviceId: string,
+  command: string,
+  params: object
+): smarthome.IntentFlow.ExecuteRequestCommands {
+  return {
+    devices: [{id: deviceId}],
+    execution: [{command, params}],
+  };
+}
+
+/**
+ * Implementation of smarthome.DataFlow.UpdResponseData for testing DeviceManager
+ */
+export class UdpResponseData implements smarthome.DataFlow.UdpResponseData {
+  constructor(requestId: string, deviceId: string) {
+    this.requestId = requestId;
+    this.deviceId = deviceId;
+  }
+  requestId: string;
+  deviceId: string;
+  protocol: smarthome.Constants.Protocol = smarthome.Constants.Protocol.UDP;
+}
